Create BeagleBoard expansion header symbol.
I've left off some of the mux'd signals otherwise the symbol would be
un-readable. If we need to use them, which I doubt we do, the symbol can be
updated. The missing signal names are for SPI channel 4 and some PWM related
stuff.

The following signal names are missing:
Pin	Name
4	GPT9_PWMEVT
6	GPT11_PWMEVT
10	GPT10_PWMEVT
12	McSPI4_SIMO
16	McSPI4_CS0
18	McSPI4_SOMI