Add custom SR 0603 footprint
The spacing between the pads of the gEDA 0603 footprint is slightly
too small to allow a trace to pass underneath the component. It also
does not contain any silk lines to show where the component should
be mounted.

This footprint has enough space between the pads to allow a single
0.2mm trace to pass through and also has two silk lines along the
edges as with the other gEDA chip footprints.

I've checked this footprint with an 0603 resistor and cap, it seems
to be fine.